Title: 실수 받아 그대로 출력, Time: -1ms, Memory: -1MB, Status: Compile Error - Codetree

+++ 250901/실수 받아 그대로 출력/enter-real-value-and-print.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main() {
     // Please write your code here.
-    int N;
+    double N;
     cout << fixed;
     cout.setprecision(2);
